pyproject.toml
[tool.poetry]
name = "portafolio"
version = "0.1.0"
description = ""
authors = ["willysan2020 <wilfredoantoniosanteliz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.11"
python-dotenv = "^1.0.0"
alembic = "^1.10.4"
pydantic = "^1.10.7"
cryptography = "^40.0.2"
bcrypt = "^4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
